+++ internal_testsuites/typescript/src/testsuite/startosis_module_test/startosis_module.test.ts
@@ -5,22 +5,38 @@ import { createEnclave } from "../../test_helpers/enclave_setup";
 import * as path from "path";
 
 const IS_PARTITIONING_ENABLED = false;
+const EMPTY_EXECUTE_PARAMS = "{}"
 
-const VALID_CASE_TEST_NAME = "startosis-module-valid";
-const VALID_KURTOSIS_MODULE_REL_PATH = "../../../../startosis/valid-kurtosis-module"
-const INVALID_KURTOSIS_MOD_TEST_NAME = "startosis-module-invalid-mod-file"
+const VALID_MODULE_WITH_TYPES_TEST_NAME = "valid-module-with-types";
+const VALID_MODULE_WITH_TYPES_REL_PATH = "../../../../startosis/valid-kurtosis-module-with-types"
+
+const VALID_MODULE_NO_TYPE_TEST_NAME = "valid-module-no-type";
+const VALID_MODULE_NO_TYPE_REL_PATH = "../../../../startosis/valid-kurtosis-module-no-type"
+
+const VALID_MODULE_NO_MODULE_INPUT_TYPE_TEST_NAME = "valid-module-no-input-type";
+const VALID_MODULE_NO_MODULE_INPUT_TYPE_REL_PATH = "../../../../startosis/valid-kurtosis-module-no-module-input-type";
+
+const INVALID_TYPES_FILE_TEST_NAME = "invalid-types-file"
+const INVALID_TYPES_FILE_REL_PATH = "../../../../startosis/invalid-types-file"
+
+const INVALID_MODULE_NO_TYPE_BUT_INPUT_ARGS_TEST_NAME = "invalid-module-no-type-input-args";
+const INVALID_MODULE_NO_TYPE_BUT_INPUT_ARGS_REL_PATH = "../../../../startosis/invalid-no-type-but-input-args";
+
+const INVALID_KURTOSIS_MOD_TEST_NAME = "invalid-module-invalid-mod-file"
 const INVALID_KURTOSIS_MOD_IN_MODULE_REL_PATH = "../../../../startosis/invalid-mod-file"
-const MISSING_MAIN_STAR_TEST_NAME = "startosis-module-no-main-file"
+
+const MISSING_MAIN_STAR_TEST_NAME = "invalid-module-no-main-file"
 const MODULE_WITH_NO_MAIN_STAR_REL_PATH = "../../../../startosis/no-main-star"
-const MISSING_MAIN_FUNCTION_TEST_NAME = "startosis-module-missing-main"
+
+const MISSING_MAIN_FUNCTION_TEST_NAME = "invalid-module-missing-main"
 const MODULE_WITH_NO_MAIN_IN_MAIN_STAR_REL_PATH = "../../../../startosis/no-main-in-main-star"
 
 
 jest.setTimeout(180000)
 
-test("Test valid startosis module execution", async () => {
+test("Test valid startosis module with types", async () => {
     // ------------------------------------- ENGINE SETUP ----------------------------------------------
-    const createEnclaveResult = await createEnclave(VALID_CASE_TEST_NAME, IS_PARTITIONING_ENABLED)
+    const createEnclaveResult = await createEnclave(VALID_MODULE_WITH_TYPES_TEST_NAME, IS_PARTITIONING_ENABLED)
 
     if(createEnclaveResult.isErr()) { throw createEnclaveResult.error }
 
@@ -28,7 +44,7 @@ test("Test valid startosis module execution", async () => {
 
     try {
         // ------------------------------------- TEST SETUP ----------------------------------------------
-        const moduleRootPath = path.join(__dirname, VALID_KURTOSIS_MODULE_REL_PATH)
+        const moduleRootPath = path.join(__dirname, VALID_MODULE_WITH_TYPES_REL_PATH)
 
         log.info(`Loading module at path '${moduleRootPath}'`)
 
@@ -64,6 +80,279 @@ test("Test valid startosis module execution", async () => {
     }
 })
 
+test("Test valid startosis module with no type", async () => {
+    // ------------------------------------- ENGINE SETUP ----------------------------------------------
+    const createEnclaveResult = await createEnclave(VALID_MODULE_NO_TYPE_TEST_NAME, IS_PARTITIONING_ENABLED)
+
+    if(createEnclaveResult.isErr()) { throw createEnclaveResult.error }
+
+    const { enclaveContext, stopEnclaveFunction } = createEnclaveResult.value
+
+    try {
+        // ------------------------------------- TEST SETUP ----------------------------------------------
+        const moduleRootPath = path.join(__dirname, VALID_MODULE_NO_TYPE_REL_PATH)
+
+        log.info(`Loading module at path '${moduleRootPath}'`)
+
+        const executeStartosisModuleResult = await enclaveContext.executeStartosisModule(moduleRootPath, EMPTY_EXECUTE_PARAMS)
+
+        if(executeStartosisModuleResult.isErr()) {
+            log.error(`An error occurred execute startosis module '${moduleRootPath}'`);
+            throw executeStartosisModuleResult.error
+        }
+        const executeStartosisModuleValue = executeStartosisModuleResult.value;
+
+        const expectedScriptOutput = "Hello World!\n"
+
+        if (expectedScriptOutput !== executeStartosisModuleValue.getSerializedScriptOutput()) {
+            throw err(new Error(`Expected output to be '${expectedScriptOutput} got '${executeStartosisModuleValue.getSerializedScriptOutput()}'`))
+        }
+
+        if (executeStartosisModuleValue.getInterpretationError() !== "") {
+            throw err(new Error(`Expected Empty Interpretation Error got '${executeStartosisModuleValue.getInterpretationError()}'`))
+        }
+
+        if (executeStartosisModuleValue.getExecutionError() !== "") {
+            throw err(new Error(`Expected Empty Execution Error got '${executeStartosisModuleValue.getExecutionError()}'`))
+        }
+
+        if (executeStartosisModuleValue.getValidationErrorsList().length != 0) {
+            throw err(new Error(`Expected Empty Validation Error got '${executeStartosisModuleValue.getValidationErrorsList()}'`))
+        }
+
+    }finally{
+        stopEnclaveFunction()
+    }
+})
+
+test("Test valid startosis module with no type - failure called with params", async () => {
+    // ------------------------------------- ENGINE SETUP ----------------------------------------------
+    const createEnclaveResult = await createEnclave(VALID_MODULE_NO_TYPE_TEST_NAME, IS_PARTITIONING_ENABLED)
+
+    if(createEnclaveResult.isErr()) { throw createEnclaveResult.error }
+
+    const { enclaveContext, stopEnclaveFunction } = createEnclaveResult.value
+
+    try {
+        // ------------------------------------- TEST SETUP ----------------------------------------------
+        const moduleRootPath = path.join(__dirname, VALID_MODULE_NO_TYPE_REL_PATH)
+
+        log.info(`Loading module at path '${moduleRootPath}'`)
+
+        const serializedParams = "{\"greetings\": \"Bonjour!\"}"
+        const executeStartosisModuleResult = await enclaveContext.executeStartosisModule(moduleRootPath, serializedParams)
+
+        if(executeStartosisModuleResult.isErr()) {
+            log.error(`An error occurred execute startosis module '${moduleRootPath}'`);
+            throw executeStartosisModuleResult.error
+        }
+        const executeStartosisModuleValue = executeStartosisModuleResult.value;
+        if (executeStartosisModuleValue.getInterpretationError() === "") {
+            throw err(new Error("Expected interpretation errors but got empty interpretation errors"))
+        }
+
+        if (!executeStartosisModuleValue.getInterpretationError().includes("A non empty parameter was passed to the module 'github.com/sample/sample-kurtosis-module' but the module doesn't contain a valid 'types.proto' file (it is either absent of invalid).")) {
+            throw err(new Error("Got interpretation error but got invalid contents"))
+        }
+
+        if (executeStartosisModuleValue.getExecutionError() !== "") {
+            throw err(new Error(`Expected Empty Execution Error got '${executeStartosisModuleValue.getExecutionError()}'`))
+        }
+
+        if (executeStartosisModuleValue.getValidationErrorsList().length != 0) {
+            throw err(new Error(`Expected Empty Validation Error got '${executeStartosisModuleValue.getValidationErrorsList()}'`))
+        }
+
+        if (executeStartosisModuleValue.getSerializedScriptOutput() != "") {
+            throw err(new Error(`Expected output to be empty got '${executeStartosisModuleValue.getSerializedScriptOutput()}'`))
+        }
+    }finally{
+        stopEnclaveFunction()
+    }
+})
+
+test("Test valid startosis module with no module input type in types file", async () => {
+    // ------------------------------------- ENGINE SETUP ----------------------------------------------
+    const createEnclaveResult = await createEnclave(VALID_MODULE_NO_MODULE_INPUT_TYPE_TEST_NAME, IS_PARTITIONING_ENABLED)
+
+    if(createEnclaveResult.isErr()) { throw createEnclaveResult.error }
+
+    const { enclaveContext, stopEnclaveFunction } = createEnclaveResult.value
+
+    try {
+        // ------------------------------------- TEST SETUP ----------------------------------------------
+        const moduleRootPath = path.join(__dirname, VALID_MODULE_NO_MODULE_INPUT_TYPE_REL_PATH)
+
+        log.info(`Loading module at path '${moduleRootPath}'`)
+
+        const executeStartosisModuleResult = await enclaveContext.executeStartosisModule(moduleRootPath, EMPTY_EXECUTE_PARAMS)
+
+        if(executeStartosisModuleResult.isErr()) {
+            log.error(`An error occurred execute startosis module '${moduleRootPath}'`);
+            throw executeStartosisModuleResult.error
+        }
+        const executeStartosisModuleValue = executeStartosisModuleResult.value;
+
+        const expectedScriptOutput = "Hello world!\n"
+
+        if (expectedScriptOutput !== executeStartosisModuleValue.getSerializedScriptOutput()) {
+            throw err(new Error(`Expected output to be '${expectedScriptOutput} got '${executeStartosisModuleValue.getSerializedScriptOutput()}'`))
+        }
+
+        if (executeStartosisModuleValue.getInterpretationError() !== "") {
+            throw err(new Error(`Expected Empty Interpretation Error got '${executeStartosisModuleValue.getInterpretationError()}'`))
+        }
+
+        if (executeStartosisModuleValue.getExecutionError() !== "") {
+            throw err(new Error(`Expected Empty Execution Error got '${executeStartosisModuleValue.getExecutionError()}'`))
+        }
+
+        if (executeStartosisModuleValue.getValidationErrorsList().length != 0) {
+            throw err(new Error(`Expected Empty Validation Error got '${executeStartosisModuleValue.getValidationErrorsList()}'`))
+        }
+
+    }finally{
+        stopEnclaveFunction()
+    }
+})
+
+test("Test valid startosis module with no module input type in types file - failure called with params", async () => {
+    // ------------------------------------- ENGINE SETUP ----------------------------------------------
+    const createEnclaveResult = await createEnclave(VALID_MODULE_NO_MODULE_INPUT_TYPE_TEST_NAME, IS_PARTITIONING_ENABLED)
+
+    if(createEnclaveResult.isErr()) { throw createEnclaveResult.error }
+
+    const { enclaveContext, stopEnclaveFunction } = createEnclaveResult.value
+
+    try {
+        // ------------------------------------- TEST SETUP ----------------------------------------------
+        const moduleRootPath = path.join(__dirname, VALID_MODULE_NO_MODULE_INPUT_TYPE_REL_PATH)
+
+        log.info(`Loading module at path '${moduleRootPath}'`)
+
+        const serializedParams = "{\"greetings\": \"Bonjour!\"}"
+        const executeStartosisModuleResult = await enclaveContext.executeStartosisModule(moduleRootPath, serializedParams)
+
+        if(executeStartosisModuleResult.isErr()) {
+            log.error(`An error occurred execute startosis module '${moduleRootPath}'`);
+            throw executeStartosisModuleResult.error
+        }
+        const executeStartosisModuleValue = executeStartosisModuleResult.value;
+        if (executeStartosisModuleValue.getInterpretationError() === "") {
+            throw err(new Error("Expected interpretation errors but got empty interpretation errors"))
+        }
+
+        if (!executeStartosisModuleValue.getInterpretationError().includes("A non empty parameter was passed to the module 'github.com/sample/sample-kurtosis-module' but 'ModuleInput' type is not defined in the module's 'types.proto' file.")) {
+            throw err(new Error("Got interpretation error but got invalid contents"))
+        }
+
+        if (executeStartosisModuleValue.getExecutionError() !== "") {
+            throw err(new Error(`Expected Empty Execution Error got '${executeStartosisModuleValue.getExecutionError()}'`))
+        }
+
+        if (executeStartosisModuleValue.getValidationErrorsList().length != 0) {
+            throw err(new Error(`Expected Empty Validation Error got '${executeStartosisModuleValue.getValidationErrorsList()}'`))
+        }
+
+        if (executeStartosisModuleValue.getSerializedScriptOutput() != "") {
+            throw err(new Error(`Expected output to be empty got '${executeStartosisModuleValue.getSerializedScriptOutput()}'`))
+        }
+    }finally{
+        stopEnclaveFunction()
+    }
+})
+
+test("Test invalid startosis module invalid types file", async () => {
+    // ------------------------------------- ENGINE SETUP ----------------------------------------------
+    const createEnclaveResult = await createEnclave(INVALID_TYPES_FILE_TEST_NAME, IS_PARTITIONING_ENABLED)
+
+    if(createEnclaveResult.isErr()) { throw createEnclaveResult.error }
+
+    const { enclaveContext, stopEnclaveFunction } = createEnclaveResult.value
+
+    try {
+        // ------------------------------------- TEST SETUP ----------------------------------------------
+        const moduleRootPath = path.join(__dirname, INVALID_TYPES_FILE_REL_PATH)
+
+        log.info(`Loading module at path '${moduleRootPath}'`)
+
+        const serializedParams = "{\"greetings\": \"Bonjour!\"}"
+        const executeStartosisModuleResult = await enclaveContext.executeStartosisModule(moduleRootPath, serializedParams)
+
+        if(executeStartosisModuleResult.isErr()) {
+            log.error(`An error occurred execute startosis module '${moduleRootPath}'`);
+            throw executeStartosisModuleResult.error
+        }
+        const executeStartosisModuleValue = executeStartosisModuleResult.value;
+        if (executeStartosisModuleValue.getInterpretationError() === "") {
+            throw err(new Error("Expected interpretation errors but got empty interpretation errors"))
+        }
+
+        if (!executeStartosisModuleValue.getInterpretationError().includes("A non empty parameter was passed to the module 'github.com/sample/sample-kurtosis-module' but the module doesn't contain a valid 'types.proto' file (it is either absent of invalid).")) {
+            throw err(new Error("Got interpretation error but got invalid contents"))
+        }
+
+        if (executeStartosisModuleValue.getExecutionError() !== "") {
+            throw err(new Error(`Expected Empty Execution Error got '${executeStartosisModuleValue.getExecutionError()}'`))
+        }
+
+        if (executeStartosisModuleValue.getValidationErrorsList().length != 0) {
+            throw err(new Error(`Expected Empty Validation Error got '${executeStartosisModuleValue.getValidationErrorsList()}'`))
+        }
+
+        if (executeStartosisModuleValue.getSerializedScriptOutput() != "") {
+            throw err(new Error(`Expected output to be empty got '${executeStartosisModuleValue.getSerializedScriptOutput()}'`))
+        }
+    }finally{
+        stopEnclaveFunction()
+    }
+})
+
+test("Test invalid startosis module no types file but input_args in main", async () => {
+    // ------------------------------------- ENGINE SETUP ----------------------------------------------
+    const createEnclaveResult = await createEnclave(INVALID_MODULE_NO_TYPE_BUT_INPUT_ARGS_TEST_NAME, IS_PARTITIONING_ENABLED)
+
+    if(createEnclaveResult.isErr()) { throw createEnclaveResult.error }
+
+    const { enclaveContext, stopEnclaveFunction } = createEnclaveResult.value
+
+    try {
+        // ------------------------------------- TEST SETUP ----------------------------------------------
+        const moduleRootPath = path.join(__dirname, INVALID_MODULE_NO_TYPE_BUT_INPUT_ARGS_REL_PATH)
+
+        log.info(`Loading module at path '${moduleRootPath}'`)
+
+        const executeStartosisModuleResult = await enclaveContext.executeStartosisModule(moduleRootPath, EMPTY_EXECUTE_PARAMS)
+
+        if(executeStartosisModuleResult.isErr()) {
+            log.error(`An error occurred execute startosis module '${moduleRootPath}'`);
+            throw executeStartosisModuleResult.error
+        }
+        const executeStartosisModuleValue = executeStartosisModuleResult.value;
+        if (executeStartosisModuleValue.getInterpretationError() === "") {
+            throw err(new Error("Expected interpretation errors but got empty interpretation errors"))
+        }
+
+        if (!executeStartosisModuleValue.getInterpretationError().includes("Evaluation error: function main missing 1 argument (input_args)")) {
+            throw err(new Error("Got interpretation error but got invalid contents"))
+        }
+
+        if (executeStartosisModuleValue.getExecutionError() !== "") {
+            throw err(new Error(`Expected Empty Execution Error got '${executeStartosisModuleValue.getExecutionError()}'`))
+        }
+
+        if (executeStartosisModuleValue.getValidationErrorsList().length != 0) {
+            throw err(new Error(`Expected Empty Validation Error got '${executeStartosisModuleValue.getValidationErrorsList()}'`))
+        }
+
+        if (executeStartosisModuleValue.getSerializedScriptOutput() != "") {
+            throw err(new Error(`Expected output to be empty got '${executeStartosisModuleValue.getSerializedScriptOutput()}'`))
+        }
+    }finally{
+        stopEnclaveFunction()
+    }
+})
+
 test("Test invalid module with invalid mod file", async () => {
     // ------------------------------------- ENGINE SETUP ----------------------------------------------
     const createEnclaveResult = await createEnclave(INVALID_KURTOSIS_MOD_TEST_NAME, IS_PARTITIONING_ENABLED)
@@ -78,7 +367,7 @@ test("Test invalid module with invalid mod file", async () => {
 
         log.info(`Loading module at path '${moduleRootPath}'`)
 
-        const executeStartosisModuleResult = await enclaveContext.executeStartosisModule(moduleRootPath, "{}")
+        const executeStartosisModuleResult = await enclaveContext.executeStartosisModule(moduleRootPath, EMPTY_EXECUTE_PARAMS)
 
         if(!executeStartosisModuleResult.isErr()) {
             throw err(new Error("Module with invalid module was expected to error but didn't"))
@@ -106,7 +395,7 @@ test("Test invalid module with no main.star", async () => {
 
         log.info(`Loading module at path '${moduleRootPath}'`)
 
-        const executeStartosisModuleResult = await enclaveContext.executeStartosisModule(moduleRootPath, "{}")
+        const executeStartosisModuleResult = await enclaveContext.executeStartosisModule(moduleRootPath, EMPTY_EXECUTE_PARAMS)
 
         if(!executeStartosisModuleResult.isErr()) {
             throw err(new Error("Module with invalid module was expected to error but didn't"))
@@ -134,7 +423,7 @@ test("Test invalid module with no main in main.star", async () => {
 
         log.info(`Loading module at path '${moduleRootPath}'`)
 
-        const executeStartosisModuleResult = await enclaveContext.executeStartosisModule(moduleRootPath, "{}")
+        const executeStartosisModuleResult = await enclaveContext.executeStartosisModule(moduleRootPath, EMPTY_EXECUTE_PARAMS)
 
         if(executeStartosisModuleResult.isErr()) {
             throw err(new Error("Unexpected execution error"))